Resolved problems with datasources

+++ repercussao/src/main/java/com/newsletter/dto/UserShareDTO.java
@@ -1,4 +1,4 @@
 package com.newsletter.dto;
 
-public record UserShareDTO(long id, double compartilhados) {
+public record UserShareDTO(long id, Integer compartilhados) {
 }
